Move tfcli error helpers to tfcli.errors package
